refactor: split http server logic
